+++ src/main/php/lang/ast/Scope.class.php
@@ -29,15 +29,6 @@ class Scope {
     $this->parent= $parent;
   }
 
-  public function define($name, $node) {
-    $definition= clone $node;
-    $definition->symbol->nud= function($self) { return $self; };
-    $definition->symbol->led= null;
-    $definition->symbol->std= null;
-    $definition->symbol->lbp= 0;
-    $this->defines[$name]= $definition;
-  }
-
   /**
    * Sets package
    *
@@ -78,9 +69,5 @@ class Scope {
     } else {
       return $name;
     }
-  }
-
-  public function find($name) {
-    return isset($this->defines[$name]) ? $this->defines[$name] : null;
   }
 }